requirements.txt
pyrogram
dotenv
Tgcrypto
psycopg2-binary